move files panel root to editor package
