+++ packages/ui/src/components/editor/layout/Panel.tsx
@@ -26,11 +26,7 @@ Ethereal Engine. All Rights Reserved.
 import React, { ReactNode } from 'react'
 import Text from '../../../primitives/tailwind/Text'
 
-const PanelIcon = ({ as: IconComponent, size = 12 }) => {
-  return <IconComponent className="mr-[6px] w-[18px] text-[var(--textColor)]" size={size} />
-}
-
-export const PanelTitle = ({ children }) => {
+export const PanelTitle = ({ children }: { children: ReactNode }) => {
   return (
     <Text fontSize="sm" className="leading-none">
       {children}
@@ -38,28 +34,6 @@ export const PanelTitle = ({ children }) => {
   )
 }
 
-export const PanelDragContainer = ({ children }) => {
-  return <div className="flex h-[30px] cursor-pointer rounded-t-md px-4 py-2">{children}</div>
+export const PanelDragContainer = ({ children }: { children: ReactNode }) => {
+  return <div className="flex h-7 cursor-pointer rounded-t-md px-4 py-2">{children}</div>
 }
-
-interface PanelProps {
-  icon?: React.ElementType
-  title: string
-  toolbarContent?: React.ReactNode
-  children?: ReactNode
-}
-
-const Panel: React.FC<PanelProps> = ({ icon, title, children, toolbarContent, ...rest }) => {
-  return (
-    <div className="bg-surface-main relative flex flex-1 select-none flex-col overflow-hidden rounded" {...rest}>
-      <div className="toolbar flex h-6 items-center border-b border-black border-opacity-20 p-1">
-        {icon && <PanelIcon as={icon} size={12} />}
-        <PanelTitle>{title}</PanelTitle>
-        {toolbarContent}
-      </div>
-      <div className="relative flex flex-1 flex-col overflow-hidden">{children}</div>
-    </div>
-  )
-}
-
-export default Panel
